# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Notebook REST API",
        "description": "Rest API for notebook version 1.0",
        "contact": {
            "email": "elk.dev.official@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/resource.json": {
            "get": {
                "responses": {
                    "200": {
                        "description": "api json resource"
                    }
                }
            }
        },
        "/api/v1/notebook": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Get list of notebooks",
                "description": "Returns list of notebooks",
                "operationId": "getNotebookList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotebookResource"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Store new notebook",
                "description": "Returns notebook data",
                "operationId": "storeNotebook",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreNotebookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotebookResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/v1/notebook/{notebook}": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Get notebook information",
                "description": "Returns notebook data",
                "operationId": "getNotebookById",
                "parameters": [
                    {
                        "name": "notebook",
                        "in": "path",
                        "description": "Notebook id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotebookResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Update existing notebook",
                "description": "Returns updated notebook data",
                "operationId": "updateNotebook",
                "parameters": [
                    {
                        "name": "notebook",
                        "in": "path",
                        "description": "Notebook id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateNotebookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotebookResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Delete existing notebook",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteNotebook",
                "parameters": [
                    {
                        "name": "notebook",
                        "in": "path",
                        "description": "Notebook id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "StoreNotebookRequest": {
                "title": "StoreNotebookRequest",
                "description": "Store Notebook request body data",
                "required": [
                    "full_name",
                    "email",
                    "phone"
                ],
                "properties": {
                    "full_name": {
                        "title": "Full name",
                        "description": "Notebook's full name",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Notebook's email",
                        "type": "string",
                        "example": "John@Doe.acme"
                    },
                    "phone": {
                        "title": "Phone",
                        "description": "Notebook's phone number",
                        "type": "string",
                        "example": "+795689632569"
                    },
                    "company": {
                        "title": "Company",
                        "description": "Notebook's company's name",
                        "type": "string",
                        "example": "Acme"
                    },
                    "birthdate": {
                        "title": "Birthdate",
                        "description": "Notebook's birthdate",
                        "type": "string",
                        "format": "YYYY-mm-dd",
                        "example": "1990-06-26"
                    },
                    "avatar": {
                        "title": "Avatar",
                        "description": "Notebook's avatar url",
                        "type": "string",
                        "example": "http://localhost:8000/public/avatar/1/avatar.jpg"
                    }
                },
                "type": "object"
            },
            "UpdateNotebookRequest": {
                "title": "UpdateNotebookRequest",
                "description": "Store Notebook request body data",
                "required": [
                    "full_name",
                    "email",
                    "phone"
                ],
                "properties": {
                    "full_name": {
                        "title": "Full name",
                        "description": "Notebook's full name",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Notebook's email",
                        "type": "string",
                        "example": "John@Doe.acme"
                    },
                    "phone": {
                        "title": "Phone",
                        "description": "Notebook's phone number",
                        "type": "string",
                        "example": "+795689632569"
                    },
                    "company": {
                        "title": "Company",
                        "description": "Notebook's company's name",
                        "type": "string",
                        "example": "Acme"
                    },
                    "birthdate": {
                        "title": "Birthdate",
                        "description": "Notebook's birthdate",
                        "type": "string",
                        "format": "YYYY-mm-dd",
                        "example": "1990-06-26"
                    },
                    "avatar": {
                        "title": "Avatar",
                        "description": "Notebook's avatar url",
                        "type": "string",
                        "example": "http://localhost:8000/public/avatar/1/avatar.jpg"
                    }
                },
                "type": "object"
            },
            "NotebookResource": {
                "title": "NotebookResource",
                "description": "Notebook resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object",
                "xml": {
                    "name": "NotebookResource"
                }
            },
            "Notebook": {
                "title": "Notebook",
                "description": "Notebook model",
                "required": [
                    "full_name",
                    "email",
                    "phone"
                ],
                "properties": {
                    "full_name": {
                        "title": "Full name",
                        "description": "Notebook's full name",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Notebook's email",
                        "type": "string",
                        "example": "John@Doe.acme"
                    },
                    "phone": {
                        "title": "Phone",
                        "description": "Notebook's phone number",
                        "type": "string",
                        "example": "+795689632569"
                    },
                    "company": {
                        "title": "Company",
                        "description": "Notebook's company's name",
                        "type": "string",
                        "example": "Acme"
                    },
                    "birthdate": {
                        "title": "Birthdate",
                        "description": "Notebook's birthdate",
                        "type": "string",
                        "format": "YYYY-mm-dd",
                        "example": "1990-06-26"
                    },
                    "avatar": {
                        "title": "Avatar",
                        "description": "Notebook's avatar url",
                        "type": "string",
                        "example": "http://localhost:8000/public/avatar/1/avatar.jpg"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Notebook"
                }
            }
        }
    },
    "tags": [
        {
            "name": "Notebook",
            "description": "API Endpoints of Notebook"
        }
    ]
}